wip: address MR review (support_recon → retarget/, split g1_planner into cli/motion/qpos utils)
