+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/g1_planner.py
@@ -13,7 +13,6 @@ Usage:
 
 from __future__ import annotations
 
-import argparse
 import os
 from pathlib import Path
 from typing import Any
@@ -25,22 +24,24 @@ from scipy.spatial.transform import Rotation, Slerp
 
 from robotic_grounding.assets import ASSET_DIR
 from robotic_grounding.motion_schema import MotionData, save_motion_parquet
+from robotic_grounding.planner.cli import parse_args
 from robotic_grounding.planner.motionbricks.inference import MotionInferenceAgent
-from robotic_grounding.planner.support_recon import (
-    reconstruct_support_for_sequence,
-)
 from robotic_grounding.planner.trajectory import build_interp_trajectory
-from robotic_grounding.planner.utils.loader import interpolate_robot_motion_data
+from robotic_grounding.planner.utils.motion import (
+    assemble_hand_contact_fields,
+    assemble_object_fields,
+    interpolate_robot_motion_data,
+)
+from robotic_grounding.planner.utils.qpos import (
+    LEFT_WRIST,
+    RIGHT_WRIST,
+    ROOT_HEIGHT,
+    build_full_qpos,
+    wrist_ee_error_from_qpos,
+)
 from robotic_grounding.planner.utils.transforms import (
     apply_local_frame_fix,
-    quat_conj,
-    quat_mul,
-    transform_contact_dir_by_part,
-    transform_contact_pos_by_part,
-    transform_primary_pos,
-    transform_primary_quat,
     transform_reference,
-    xyzw_to_wxyz,
 )
 from robotic_grounding.planner.utils.validation import (
     assert_motion_parquet_invariants,
@@ -49,6 +50,9 @@ from robotic_grounding.planner.utils.validation import (
 )
 from robotic_grounding.planner.visualization import visualize
 from robotic_grounding.retarget.data_logger import ManoDex3Data, ManoSharpaData
+from robotic_grounding.retarget.support_recon import (
+    reconstruct_support_for_sequence,
+)
 
 # -- Constants ------------------------------------------------------------------
 
@@ -56,8 +60,6 @@ FPS = 30
 HOLD_START_S = 5.0
 INTERP_DURATION_S = 5.0
 HOLD_END_S = 5.0
-ROOT_HEIGHT = 0.793
-ROOT_FIX_COMPONENTS = ("x", "y", "z", "roll", "pitch", "yaw")
 
 NOMINAL_JOINTS = {
     22: -0.5,  # left_shoulder_pitch
@@ -67,24 +69,6 @@ NOMINAL_JOINTS = {
     30: -0.2,  # right_shoulder_roll
     32: 0.0,  # right_elbow
 }
-
-LEG_OVERRIDES = {
-    "left_hip_pitch_joint": -0.1,
-    "left_hip_roll_joint": 0.0,
-    "left_hip_yaw_joint": 0.0,
-    "left_knee_joint": 0.4,
-    "left_ankle_pitch_joint": -0.2,
-    "left_ankle_roll_joint": 0.0,
-    "right_hip_pitch_joint": -0.1,
-    "right_hip_roll_joint": 0.0,
-    "right_hip_yaw_joint": 0.0,
-    "right_knee_joint": 0.4,
-    "right_ankle_pitch_joint": -0.2,
-    "right_ankle_roll_joint": 0.0,
-}
-
-LEFT_WRIST = "left_wrist_yaw_link"
-RIGHT_WRIST = "right_wrist_yaw_link"
 
 _ASSETS_DIR = Path(__file__).parent / "assets"
 
@@ -293,184 +277,6 @@ def load_v2p_reference(
     }
 
 
-G1_BODY_JOINT_NAMES = [
-    "left_hip_pitch_joint",
-    "left_hip_roll_joint",
-    "left_hip_yaw_joint",
-    "left_knee_joint",
-    "left_ankle_pitch_joint",
-    "left_ankle_roll_joint",
-    "right_hip_pitch_joint",
-    "right_hip_roll_joint",
-    "right_hip_yaw_joint",
-    "right_knee_joint",
-    "right_ankle_pitch_joint",
-    "right_ankle_roll_joint",
-    "waist_yaw_joint",
-    "waist_roll_joint",
-    "waist_pitch_joint",
-    "left_shoulder_pitch_joint",
-    "left_shoulder_roll_joint",
-    "left_shoulder_yaw_joint",
-    "left_elbow_joint",
-    "left_wrist_roll_joint",
-    "left_wrist_pitch_joint",
-    "left_wrist_yaw_joint",
-    "right_shoulder_pitch_joint",
-    "right_shoulder_roll_joint",
-    "right_shoulder_yaw_joint",
-    "right_elbow_joint",
-    "right_wrist_roll_joint",
-    "right_wrist_pitch_joint",
-    "right_wrist_yaw_joint",
-]
-
-
-def build_body_joint_mapping(model: mujoco.MjModel) -> dict[int, int]:
-    """Map 29 body DOFs to combined model qpos indices."""
-    mapping = {}
-    for dof_idx, jname in enumerate(G1_BODY_JOINT_NAMES):
-        jid = mujoco.mj_name2id(model, mujoco.mjtObj.mjOBJ_JOINT, jname)
-        if jid >= 0:
-            mapping[dof_idx] = int(model.jnt_qposadr[jid])
-    return mapping
-
-
-def build_finger_mapping(model: mujoco.MjModel, joint_names: list[str]) -> list[int]:
-    """Map finger joint names to model qpos indices."""
-    result = []
-    for jname in joint_names:
-        try:
-            result.append(int(model.jnt_qposadr[model.joint(jname).id]))
-        except Exception:
-            result.append(-1)
-    return result
-
-
-def _wrist_ee_error_from_qpos(
-    full_qpos: np.ndarray,
-    ref_data: dict,
-    model: mujoco.MjModel,
-) -> float:
-    """Mean L2 distance from FK'd wrist_yaw_link bodies to V2P wrist targets.
-
-    Used to score heading-offset candidates during the local search.
-    """
-    data = mujoco.MjData(model)
-    li = model.body(LEFT_WRIST).id
-    ri = model.body(RIGHT_WRIST).id
-    target_l = np.asarray(ref_data["left_pos"], dtype=np.float64)
-    target_r = np.asarray(ref_data["right_pos"], dtype=np.float64)
-    T = min(full_qpos.shape[0], target_l.shape[0], target_r.shape[0])
-    err_l = np.empty(T, dtype=np.float64)
-    err_r = np.empty(T, dtype=np.float64)
-    for t in range(T):
-        data.qpos[: full_qpos.shape[1]] = full_qpos[t]
-        mujoco.mj_forward(model, data)
-        err_l[t] = np.linalg.norm(data.xpos[li] - target_l[t])
-        err_r[t] = np.linalg.norm(data.xpos[ri] - target_r[t])
-    return float(((err_l + err_r) / 2.0).mean())
-
-
-def _root_fix_component_set(
-    components: tuple[str, ...] | list[str] = (),
-    *,
-    fix_root_pos: bool = False,
-    fix_root_rot: bool = False,
-    fix_root_z: bool = False,
-    fix_root_rp: bool = False,
-) -> set[str]:
-    """Merge the generalized root component list with legacy root flags."""
-    result = set(components or ())
-    invalid = result.difference(ROOT_FIX_COMPONENTS)
-    if invalid:
-        raise ValueError(f"Unknown root fix component(s): {sorted(invalid)}")
-    if fix_root_pos:
-        result.update(("x", "y", "z"))
-    if fix_root_z:
-        result.add("z")
-    if fix_root_rot:
-        result.update(("roll", "pitch", "yaw"))
-    if fix_root_rp:
-        result.update(("roll", "pitch"))
-    return result
-
-
-def _root_wxyz_with_fixed_components(
-    q_xyzw: np.ndarray, fixed_components: set[str]
-) -> np.ndarray:
-    """Apply roll/pitch/yaw root clamps while preserving free components."""
-    if not fixed_components.intersection(("roll", "pitch", "yaw")):
-        return xyzw_to_wxyz(q_xyzw)
-
-    euler_xyz = Rotation.from_quat(q_xyzw).as_euler("xyz", degrees=False)
-    if "roll" in fixed_components:
-        euler_xyz[0] = 0.0
-    if "pitch" in fixed_components:
-        euler_xyz[1] = 0.0
-    if "yaw" in fixed_components:
-        euler_xyz[2] = 0.0
-    q_fixed_xyzw = Rotation.from_euler("xyz", euler_xyz).as_quat()
-    return xyzw_to_wxyz(q_fixed_xyzw).astype(np.float32)
-
-
-def build_full_qpos(
-    planned_qpos: np.ndarray,
-    ref_data: dict[str, Any],
-    model: mujoco.MjModel,
-    T_save: int,
-    fix_lower_body: bool = False,
-    fix_root_pos: bool = False,
-    fix_root_rot: bool = False,
-    fix_root_z: bool = False,
-    fix_root_rp: bool = False,
-    fix_root_components: tuple[str, ...] | list[str] = (),
-) -> tuple[np.ndarray, dict[int, int], list[int], list[int]]:
-    """Combine planned body + reference fingers + (optionally fixed) parts."""
-    nq = model.nq
-    full_qpos = np.zeros((T_save, nq), dtype=np.float32)
-    body_map = build_body_joint_mapping(model)
-    l_finger_map = build_finger_mapping(model, ref_data.get("left_joint_names", []))
-    r_finger_map = build_finger_mapping(model, ref_data.get("right_joint_names", []))
-    fixed_root = _root_fix_component_set(
-        fix_root_components,
-        fix_root_pos=fix_root_pos,
-        fix_root_rot=fix_root_rot,
-        fix_root_z=fix_root_z,
-        fix_root_rp=fix_root_rp,
-    )
-
-    for t in range(T_save):
-        t_plan = min(t, planned_qpos.shape[0] - 1)
-        full_qpos[t, :3] = planned_qpos[t_plan, :3]
-        if "x" in fixed_root:
-            full_qpos[t, 0] = 0.0
-        if "y" in fixed_root:
-            full_qpos[t, 1] = 0.0
-        if "z" in fixed_root:
-            full_qpos[t, 2] = ROOT_HEIGHT
-        full_qpos[t, 3:7] = _root_wxyz_with_fixed_components(
-            planned_qpos[t_plan, 3:7], fixed_root
-        )
-        for dof_idx, qi in body_map.items():
-            full_qpos[t, qi] = planned_qpos[t_plan, 7 + dof_idx]
-        if fix_lower_body:
-            for jname, val in LEG_OVERRIDES.items():
-                try:
-                    full_qpos[t, int(model.jnt_qposadr[model.joint(jname).id])] = val
-                except Exception:
-                    pass
-        f_idx = min(t, ref_data["left_finger_joints"].shape[0] - 1)
-        for j, qi in enumerate(l_finger_map):
-            if qi >= 0 and j < ref_data["left_finger_joints"].shape[1]:
-                full_qpos[t, qi] = ref_data["left_finger_joints"][f_idx, j]
-        for j, qi in enumerate(r_finger_map):
-            if qi >= 0 and j < ref_data["right_finger_joints"].shape[1]:
-                full_qpos[t, qi] = ref_data["right_finger_joints"][f_idx, j]
-
-    return full_qpos, body_map, l_finger_map, r_finger_map
-
-
 def save_planner_parquet(
     output_dir: str,
     full_qpos: np.ndarray,
@@ -533,265 +339,23 @@ def save_planner_parquet(
     object_body_position = obj_body_pos_arr.tolist()
     object_body_wxyz = obj_body_wxyz_arr.tolist()
 
-    # object_root_* mirrors body 0 of the planner-frame object pose.
-    # The env reads object_root_position[0] for the articulated scene init
-    # pose; deriving from body 0 keeps that init aligned with where the
-    # trajectory actually starts, regardless of whether the upstream motion
-    # carried a separately-resampled root field.
-    object_root_position: list = obj_body_pos_arr[:, 0, :].astype(np.float32).tolist()
-    object_root_axis_angle: list = (
-        Rotation.from_quat(obj_body_wxyz_arr[:, 0, :], scalar_first=True)
-        .as_rotvec()
-        .astype(np.float32)
-        .tolist()
+    motion = ref_raw.get("_motion_data")
+    object_fields = assemble_object_fields(
+        motion=motion,
+        obj_body_pos_arr=obj_body_pos_arr,
+        obj_body_wxyz_arr=obj_body_wxyz_arr,
+        T_use=T_use,
+        fallback_object_name=str(ref_data.get("object_name", "box")),
+        resolve_asset_path=_resolve_asset_path_for_output,
+        warn_missing_deps=warn_missing_urdf_mesh_deps,
     )
 
-    # Object metadata carried over from the upstream ManoSharpaData file.
-    motion = ref_raw.get("_motion_data")
-    object_name = str(ref_data.get("object_name", "box"))
-    object_body_names = ["object"]
-    safe_object_body_names = ["object"]
-    object_mesh_paths: list[str] = []
-    object_urdf_paths: list[str] = []
-    object_mesh_radius: list[float] | None = None
-    object_articulation: list[float] = [0.0] * T_use
-    safe_object_name = object_name
-    if motion is not None:
-        if getattr(motion, "object_body_names", None):
-            object_body_names = list(motion.object_body_names)
-        if getattr(motion, "safe_object_body_names", None):
-            safe_object_body_names = list(motion.safe_object_body_names)
-        if getattr(motion, "object_mesh_paths", None):
-            object_mesh_paths = [
-                str(resolved)
-                for p in motion.object_mesh_paths
-                if (resolved := _resolve_asset_path_for_output(p))
-            ]
-        if getattr(motion, "object_urdf_paths", None):
-            object_urdf_paths = [
-                str(resolved)
-                for p in motion.object_urdf_paths
-                if (resolved := _resolve_asset_path_for_output(p))
-            ]
-            # The URDF can resolve locally while its <mesh filename=> visual
-            # or collision dependencies don't. Surface those gaps now so the
-            # user fixes the workspace before training hits the import crash.
-            warn_missing_urdf_mesh_deps(object_urdf_paths)
-        if getattr(motion, "object_mesh_radius", None):
-            object_mesh_radius = [float(r) for r in motion.object_mesh_radius]
-        if getattr(motion, "safe_object_name", None):
-            safe_object_name = motion.safe_object_name
-        obj_art = getattr(motion, "object_articulation", None)
-        if obj_art is not None:
-            object_articulation = np.asarray(obj_art, dtype=np.float32)[:T_use].tolist()
-
-    # Per-side hand groups from the V2P retargeting.
-    hand_sides: list[str] = []
-    hand_frame_names: list[list[str]] = []
-    hand_frames_w: list[list[list[list[float]]]] = []
-    hand_finger_joint_names: list[list[str]] = []
-    hand_finger_joints: list[list[list[float]]] = []
-    # Per-side contact groups.
-    hand_link_contact_positions: list[list[list[list[float]]]] = []
-    hand_link_contact_normals: list[list[list[list[float]]]] = []
-    hand_object_contact_positions: list[list[list[list[float]]]] = []
-    hand_object_contact_normals: list[list[list[list[float]]]] = []
-    hand_object_contact_part_ids: list[list[list[int]]] = []
-    hand_contact_link_names: list[list[str]] = []
-    hand_contact_active: list[list[float]] = []
-
-    if motion is not None:
-        # Build the per-frame V2P→planner rigid transform anchored on the
-        # primary object body. transform_reference already applied this same
-        # transform to ee_pose_w / object_body_position (the planner-frame
-        # arrays in ref_data). Applying it here lands hand_frames_w and the
-        # contact arrays in the same frame, so the env can compare them
-        # against the planner-frame object pose without a frame mix.
-        raw_obj_pos_all = np.asarray(motion.object_body_position, dtype=np.float32)[
-            :T_use
-        ]
-        raw_obj_quat_all = np.asarray(motion.object_body_wxyz, dtype=np.float32)[:T_use]
-        if raw_obj_pos_all.ndim == 2:
-            raw_obj_pos_all = raw_obj_pos_all[:, None]
-            raw_obj_quat_all = raw_obj_quat_all[:, None]
-        dst_obj_pos_all = obj_body_pos_arr
-        dst_obj_quat_all = obj_body_wxyz_arr
-        # Length-align in case `_trim_motion_data_range` skipped a field whose
-        # length didn't match the wrist trim — slicing both to a common T
-        # keeps `_transform_*` invocations broadcast-safe.
-        common_T = min(
-            raw_obj_pos_all.shape[0],
-            dst_obj_pos_all.shape[0],
-        )
-        raw_obj_pos_all = raw_obj_pos_all[:common_T]
-        raw_obj_quat_all = raw_obj_quat_all[:common_T]
-        dst_obj_pos_all_aligned = dst_obj_pos_all[:common_T]
-        dst_obj_quat_all_aligned = dst_obj_quat_all[:common_T]
-        primary_r_rel = quat_mul(
-            dst_obj_quat_all_aligned[:, 0], quat_conj(raw_obj_quat_all[:, 0])
-        )
-        raw_primary_pos = raw_obj_pos_all[:, 0]
-        dst_primary_pos = dst_obj_pos_all_aligned[:, 0]
-
-        for side in ("left", "right"):
-            wrist_pos = getattr(motion, f"robot_{side}_wrist_position", None)
-            if wrist_pos is None:
-                continue
-            hand_sides.append(side)
-            frames = getattr(motion, f"robot_{side}_frames", None) or []
-            frame_names = getattr(motion, f"{side}_robot_frame_names", None) or []
-            finger_joints = getattr(motion, f"robot_{side}_finger_joints", None) or []
-            finger_joint_names = (
-                getattr(motion, f"{side}_robot_finger_joint_names", None) or []
-            )
-            link_contacts = (
-                getattr(motion, f"mano_{side}_link_contact_positions", None) or []
-            )
-            link_normals = (
-                getattr(motion, f"mano_{side}_link_contact_normals", None) or []
-            )
-            obj_contacts = (
-                getattr(motion, f"mano_{side}_object_contact_positions", None) or []
-            )
-            obj_normals = (
-                getattr(motion, f"mano_{side}_object_contact_normals", None) or []
-            )
-            part_ids_attr = (
-                getattr(motion, f"mano_{side}_object_contact_part_ids", None) or []
-            )
-
-            hand_frame_names.append(list(frame_names))
-            # Lift hand_frames_w into the planner frame. The consumer in
-            # tracking_command._precompute_hand_keypoints_in_object_frame
-            # combines these keypoint poses with object_body_position to build
-            # the wrist/fingertip targets; passing through V2P-frame keypoints
-            # against a planner-frame object pose silently produces targets up
-            # to a metre off.
-            if frames:
-                frames_arr = np.asarray(frames, dtype=np.float32)[:common_T]
-                frame_pos = transform_primary_pos(
-                    frames_arr[..., :3],
-                    raw_primary_pos,
-                    dst_primary_pos,
-                    primary_r_rel,
-                )
-                frame_quat = transform_primary_quat(frames_arr[..., 3:], primary_r_rel)
-                hand_frames_w.append(
-                    np.concatenate([frame_pos, frame_quat], axis=-1).tolist()
-                )
-            else:
-                hand_frames_w.append([])
-
-            hand_finger_joint_names.append(list(finger_joint_names))
-            hand_finger_joints.append(
-                np.asarray(finger_joints, dtype=np.float32)[:T_use].tolist()
-                if finger_joints
-                else []
-            )
-            hand_contact_link_names.append([])
-
-            # part_ids are 1-indexed object-body indices that drive the
-            # per-body contact transform. Some loaders carry them in a
-            # dedicated array (often left at source fps), others embed them
-            # in the 4th column of the contact-position arrays (already at
-            # planner fps). Probe both and nearest-neighbor upsample the
-            # dedicated array if it hasn't been interpolated.
-            part_ids_arr: np.ndarray | None = None
-            if len(part_ids_attr):
-                src = np.asarray(part_ids_attr, dtype=np.int64)
-                if src.shape[0] >= common_T:
-                    part_ids_arr = src[:common_T]
-                elif src.shape[0] > 0:
-                    src_t = np.linspace(0.0, 1.0, src.shape[0])
-                    dst_t = np.linspace(0.0, 1.0, common_T)
-                    nn_idx = np.clip(
-                        np.searchsorted(src_t, dst_t, side="right") - 1,
-                        0,
-                        src.shape[0] - 1,
-                    )
-                    part_ids_arr = src[nn_idx]
-            if part_ids_arr is None and obj_contacts:
-                oc_probe = np.asarray(obj_contacts, dtype=np.float32)
-                if oc_probe.ndim == 3 and oc_probe.shape[-1] >= 4:
-                    part_ids_arr = np.rint(oc_probe[:common_T, :, 3]).astype(np.int64)
-                    inactive = (
-                        np.linalg.norm(oc_probe[:common_T, :, :3], axis=-1) <= 1e-8
-                    )
-                    part_ids_arr = np.where(inactive, 0, part_ids_arr)
-
-            # Contacts ride the same per-body rigid transform as the object
-            # bodies they reference. Positions translate + rotate, normals
-            # rotate only.
-            if obj_contacts:
-                oc_a = np.asarray(obj_contacts, dtype=np.float32)[:common_T, :, :3]
-                oc_transformed = transform_contact_pos_by_part(
-                    oc_a,
-                    raw_obj_pos_all,
-                    dst_obj_pos_all_aligned,
-                    raw_obj_quat_all,
-                    dst_obj_quat_all_aligned,
-                    part_ids_arr,
-                )
-                hand_object_contact_positions.append(oc_transformed.tolist())
-            else:
-                hand_object_contact_positions.append([])
-
-            if obj_normals:
-                on_a = np.asarray(obj_normals, dtype=np.float32)[:common_T, :, :3]
-                on_transformed = transform_contact_dir_by_part(
-                    on_a,
-                    raw_obj_quat_all,
-                    dst_obj_quat_all_aligned,
-                    part_ids_arr,
-                )
-                hand_object_contact_normals.append(on_transformed.tolist())
-            else:
-                hand_object_contact_normals.append([])
-
-            if link_contacts:
-                lc_a = np.asarray(link_contacts, dtype=np.float32)[:common_T, :, :3]
-                lc_transformed = transform_contact_pos_by_part(
-                    lc_a,
-                    raw_obj_pos_all,
-                    dst_obj_pos_all_aligned,
-                    raw_obj_quat_all,
-                    dst_obj_quat_all_aligned,
-                    part_ids_arr,
-                )
-                hand_link_contact_positions.append(lc_transformed.tolist())
-            else:
-                hand_link_contact_positions.append([])
-
-            if link_normals:
-                ln_a = np.asarray(link_normals, dtype=np.float32)[:common_T, :, :3]
-                ln_transformed = transform_contact_dir_by_part(
-                    ln_a,
-                    raw_obj_quat_all,
-                    dst_obj_quat_all_aligned,
-                    part_ids_arr,
-                )
-                hand_link_contact_normals.append(ln_transformed.tolist())
-            else:
-                hand_link_contact_normals.append([])
-
-            hand_object_contact_part_ids.append(
-                part_ids_arr.tolist() if part_ids_arr is not None else []
-            )
-
-            # Per-frame contact-active mask: 1 when at least one contact point
-            # is recorded against the object, 0 otherwise. Derived from the
-            # already-upsampled `obj_contacts` so the mask length matches the
-            # other per-frame contact arrays. tracking_command refuses to load
-            # motion files where both sides are absent, so always emit a
-            # per-side mask (zero-filled in the worst case).
-            if obj_contacts:
-                cp = np.asarray(obj_contacts, dtype=np.float32)[:common_T, :, :3]
-                active = (
-                    (np.abs(cp).sum(axis=-1) > 1e-5).any(axis=-1).astype(np.float32)
-                )
-            else:
-                active = np.zeros((common_T,), dtype=np.float32)
-            hand_contact_active.append(active.tolist())
+    hand_fields = assemble_hand_contact_fields(
+        motion=motion,
+        obj_body_pos_arr=obj_body_pos_arr,
+        obj_body_wxyz_arr=obj_body_wxyz_arr,
+        T_use=T_use,
+    )
 
     robot_name = "g1" if robot_type == "sharpa" else "g1_dex3"
     # ee_link_names tells the env which body the EE pose was recorded from.
@@ -816,30 +380,10 @@ def save_planner_parquet(
         robot_joint_positions=robot_joint_positions,
         ee_link_names=ee_link_names,
         ee_pose_w=ee_pose_w.tolist(),
-        object_name=object_name,
-        safe_object_name=safe_object_name,
-        object_body_names=object_body_names,
-        safe_object_body_names=safe_object_body_names,
-        object_mesh_paths=object_mesh_paths,
-        object_urdf_paths=object_urdf_paths,
-        object_mesh_radius=object_mesh_radius,
-        object_articulation=object_articulation,
-        object_root_axis_angle=object_root_axis_angle,
-        object_root_position=object_root_position,
         object_body_position=object_body_position,
         object_body_wxyz=object_body_wxyz,
-        hand_sides=hand_sides,
-        hand_frame_names=hand_frame_names,
-        hand_frames_w=hand_frames_w,
-        hand_finger_joint_names=hand_finger_joint_names,
-        hand_finger_joints=hand_finger_joints,
-        hand_contact_link_names=hand_contact_link_names,
-        hand_link_contact_positions=hand_link_contact_positions,
-        hand_link_contact_normals=hand_link_contact_normals,
-        hand_object_contact_positions=hand_object_contact_positions,
-        hand_object_contact_normals=hand_object_contact_normals,
-        hand_object_contact_part_ids=hand_object_contact_part_ids,
-        hand_contact_active=hand_contact_active,
+        **object_fields,
+        **hand_fields,
     )
     # Layout: `<output_dir>/planner_processed/sequence_id=…/robot_name=…/*.parquet`
     # with the support USD as a sibling at `<output_dir>/reconstructed_stage/`.
@@ -853,139 +397,6 @@ def save_planner_parquet(
     # corruption can't make it past planning.
     assert_motion_parquet_invariants(partition_dir, robot_type=robot_type)
     return T_use
-
-
-# -- CLI -----------------------------------------------------------------------
-
-
-def parse_args() -> argparse.Namespace:
-    """Build the planner CLI parser and return the parsed namespace."""
-    parser = argparse.ArgumentParser(description="G1 whole-body planner")
-    parser.add_argument("--robot", choices=["sharpa", "dex3"], default="sharpa")
-    parser.add_argument("--v2p_parquet", required=True)
-    parser.add_argument("--v2p_robot_name", default="sharpa_wave")
-    parser.add_argument("--v2p_sequence", default="box_grab")
-    parser.add_argument("--v2p_trajectory_id", type=int, default=0)
-    parser.add_argument(
-        "--v2p_start_frame",
-        type=int,
-        default=0,
-        help=(
-            "Drop this many frames from the interpolated V2P reference before "
-            "building the planner warmup/interp trajectory. Useful for skipping "
-            "dataset-specific T-pose/approach lead-ins."
-        ),
-    )
-    parser.add_argument(
-        "--v2p_start_at_first_contact",
-        action="store_true",
-        help=(
-            "Start the reference at the first detected hand-object contact "
-            "minus --v2p_pre_contact_frames."
-        ),
-    )
-    parser.add_argument(
-        "--v2p_pre_contact_frames",
-        type=int,
-        default=10,
-        help="Number of interpolated V2P frames to keep before first contact.",
-    )
-    parser.add_argument(
-        "--v2p_end_after_last_contact_frames",
-        type=int,
-        default=-1,
-        help=(
-            "If >= 0, truncate the interpolated V2P reference after the last "
-            "detected hand-object contact plus this many frames. A value of 0 "
-            "keeps through the last contact frame."
-        ),
-    )
-    parser.add_argument("--target_fps", type=float, default=150.0)
-    parser.add_argument("--hold_start_s", type=float, default=HOLD_START_S)
-    parser.add_argument("--interp_s", type=float, default=INTERP_DURATION_S)
-    parser.add_argument("--hold_end_s", type=float, default=HOLD_END_S)
-    parser.add_argument(
-        "--no_approach",
-        action="store_true",
-        help=(
-            "Disable the planner's nominal hold/interp/hold approach segment. "
-            "The generated trajectory starts directly at the V2P reference."
-        ),
-    )
-    parser.add_argument(
-        "--workspace_offset", type=float, nargs=3, default=[-0.10, 0.0, -0.15]
-    )
-    parser.add_argument("--ref_seconds", type=float, default=-1)
-    parser.add_argument("--output", default=None)
-    parser.add_argument("--no_viewer", action="store_true")
-    parser.add_argument("--ik_verify", action="store_true")
-    parser.add_argument("--ik_plan", action="store_true")
-    parser.add_argument(
-        "--fix_lower_body",
-        action="store_true",
-        help=(
-            "Override the model's lower-body (hip/knee/ankle) predictions "
-            "with a static crouch and run the AR-aware loop that pins those "
-            "bodies in the model's chunk seeds."
-        ),
-    )
-    parser.add_argument(
-        "--fix_root",
-        nargs="+",
-        choices=ROOT_FIX_COMPONENTS,
-        default=(),
-        help=(
-            "Pin selected root components. Components are x y z roll pitch yaw; "
-            "e.g. '--fix_root z roll pitch' clamps height and roll/pitch while "
-            "leaving root XY translation and yaw free."
-        ),
-    )
-    parser.add_argument(
-        "--fix_root_pos",
-        action="store_true",
-        help="Legacy alias for '--fix_root x y z'.",
-    )
-    parser.add_argument(
-        "--fix_root_z",
-        action="store_true",
-        help="Legacy alias for '--fix_root z'.",
-    )
-    parser.add_argument(
-        "--fix_root_rot",
-        action="store_true",
-        help="Legacy alias for '--fix_root roll pitch yaw'.",
-    )
-    parser.add_argument(
-        "--fix_root_rp",
-        action="store_true",
-        help="Legacy alias for '--fix_root roll pitch'.",
-    )
-    parser.add_argument(
-        "--no_smooth_qpos",
-        action="store_true",
-        help="Disable post-inference qpos smoothing (global Hamming + boundary blend).",
-    )
-    parser.add_argument(
-        "--search_heading_deg",
-        type=float,
-        default=0.0,
-        help=(
-            "If > 0, run inference at heading offsets [-N, -N/2, 0, +N/2, +N] "
-            "degrees around the heading-toward-object correction and pick the "
-            "candidate with the lowest mean wrist tracking error."
-        ),
-    )
-    parser.add_argument(
-        "--heading_align_frame",
-        choices=("start", "first_contact"),
-        default="start",
-        help=(
-            "Frame used for the heading-toward-object correction. 'start' "
-            "keeps the legacy behavior; 'first_contact' uses the detected "
-            "first contact frame within the trimmed reference."
-        ),
-    )
-    return parser.parse_args()
 
 
 def main() -> None:
@@ -1178,7 +589,7 @@ def main() -> None:
             fix_root_rp=args.fix_root_rp,
             fix_root_components=args.fix_root,
         )
-        score = _wrist_ee_error_from_qpos(full_qpos_local, ref_data_local, model)
+        score = wrist_ee_error_from_qpos(full_qpos_local, ref_data_local, model)
         return (
             ref_data_local,
             (traj_lp_l, traj_lq_l, traj_rp_l, traj_rq_l),
